uartConnector/pcPart/pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"
[project]
name = "uart6502"
version = "0.6.0"
authors = [
    { name="Strannik", email="strannik.anton@gmail.com" },
]
description = "Communicate with 6502 via UART"
readme = "README.md"
requires-python = ">=3.7"
classifiers = [
    "Programming Language :: Python :: 3",
    "Operating System :: OS Independent",
]
dependencies = [
    "pyserial",
]
[project.scripts]
uart6502 = "uart6502.main:run"
[tool.hatch.build.targets.sdist]
exclude = [
    "/.venv",
    ".idea",
    ".gitignore",
    "devRun.py"
]
